requirements.txt
anyio==3.3.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.0.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.2
django-decouple==2.1
django-redis==4.12.1
djangorestframework==3.12.2
entrypoints==0.3
fastapi==0.70.0
filelock==3.0.12
future==0.18.2
gpg===1.13.1-unknown
gunicorn==20.0.4
h11==0.12.0
httplib2==0.14.0
httptools==0.2.0
idna==2.8
ipykernel==6.1.0
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.21
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.3
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandas==1.1.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pwnedpasswords==2.0.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydantic==1.8.2
pygame==2.0.0
Pygments==2.9.0
PyGObject==3.36.0
PyInquirer==1.0.3
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.2.1
redis==3.5.3
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
ssh-import-id==5.10
starlette==0.16.0
systemd-python==234
terminado==0.11.0
testpath==0.5.0
toml==0.10.2
torbrowser-launcher==0.3.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.0.34
wadllib==1.3.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.0
wrapt==1.12.1
xkit==0.0.0
